Add test package (#11)
* Add test package with some compilation fail checks

* Update CI

* Run fmt and update workflow

* Get workflow to run all tests in workspace
