+++ .github/workflows/rust.yml
@@ -9,7 +9,19 @@ on:
 env:
   CARGO_TERM_COLOR: always
 
+
+
 jobs:
+  lint:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Check formatting
+      run: cargo fmt --check  --verbose
+    - name: Run Clippy
+      run: cargo clippy --verbose -- -D warnings
+
   test:
     runs-on: ubuntu-latest
 
@@ -17,9 +29,5 @@ jobs:
     - uses: actions/checkout@v3
     - name: Build
       run: cargo build --verbose
-    - name: Run tests
-      run: cargo test --verbose
-    - name: Check formatting
-      run: cargo fmt --check  --verbose
-    - name: Run Clippy
-      run: cargo clippy --verbose -- -D warnings
+    - name: Run all tests in workspace
+      run: cargo test --workspace --verbose
