+++ const_typed_builder_test/compile_fail/mixed_mandatory.stderr
@@ -1,0 +1,18 @@
+error[E0599]: no method named `build` found for struct `FooBuilder<true, false>` in the current scope
+  --> ./compile_fail/mixed_mandatory.rs:16:10
+   |
+4  |       #[derive(Debug, Default, PartialEq, Eq, Builder)]
+   |                                               ------- method `build` not found for this struct
+...
+12 |       let foo = Foo::builder()
+   |  _______________-
+13 | |         .bar("Hello".to_string())
+14 | |         .qux(Some("Hello".to_string()))
+15 | |         .quz(Some("world!".to_string()))
+16 | |         .build();
+   | |         -^^^^^ method not found in `FooBuilder<true, false>`
+   | |_________|
+   |
+   |
+   = note: the method was found for
+           - `FooBuilder<true, true>`
