+++ const_typed_builder_test/compile_fail/multiple_mandatory.rs
@@ -1,0 +1,17 @@
+use const_typed_builder::Builder;
+
+fn main() {
+    #[derive(Debug, Default, PartialEq, Eq, Builder)]
+    pub struct Foo {
+        bar: String,
+        baz: String,
+        qux: String,
+        quz: String,
+    }
+
+    let foo = Foo::builder()
+        .bar("Hello".to_string())
+        .baz("world!".to_string())
+        .qux("Hello".to_string())
+        .build();
+}
