+++ const_typed_builder_test/compile_fail/option_explicit.rs
@@ -1,0 +1,10 @@
+use const_typed_builder::Builder;
+
+fn main() {
+    #[derive(Debug, Default, PartialEq, Eq, Builder)]
+    pub struct Foo {
+        bar: std::option::Option<String>,
+    }
+
+    let foo = Foo::builder().bar("Hello world!".to_string());
+}
