+++ const_typed_builder_test/compile_fail/option_explicit.stderr
@@ -1,0 +1,19 @@
+error[E0308]: mismatched types
+ --> ./compile_fail/option_explicit.rs:9:34
+  |
+9 |     let foo = Foo::builder().bar("Hello world!".to_string());
+  |                              --- ^^^^^^^^^^^^^^^^^^^^^^^^^^ expected `Option<String>`, found `String`
+  |                              |
+  |                              arguments to this method are incorrect
+  |
+  = note: expected enum `Option<String>`
+           found struct `String`
+note: method defined here
+ --> ./compile_fail/option_explicit.rs:6:9
+  |
+6 |         bar: std::option::Option<String>,
+  |         ^^^-----------------------------
+help: try wrapping the expression in `Some`
+  |
+9 |     let foo = Foo::builder().bar(Some("Hello world!".to_string()));
+  |                                  +++++                          +
