+++ const_typed_builder_test/compile_fail/option_not_mandatory.stderr
@@ -1,0 +1,19 @@
+error[E0308]: mismatched types
+ --> ./compile_fail/option_not_mandatory.rs:8:34
+  |
+8 |     let foo = Foo::builder().bar("Hello world!".to_string()).build();
+  |                              --- ^^^^^^^^^^^^^^^^^^^^^^^^^^ expected `Option<String>`, found `String`
+  |                              |
+  |                              arguments to this method are incorrect
+  |
+  = note: expected enum `Option<String>`
+           found struct `String`
+note: method defined here
+ --> ./compile_fail/option_not_mandatory.rs:6:9
+  |
+6 |         bar: Option<String>,
+  |         ^^^----------------
+help: try wrapping the expression in `Some`
+  |
+8 |     let foo = Foo::builder().bar(Some("Hello world!".to_string())).build();
+  |                                  +++++                          +
