+++ const_typed_builder_test/compile_fail/optional_mandatory.rs
@@ -1,0 +1,11 @@
+use const_typed_builder::Builder;
+
+fn main() {
+    #[derive(Debug, Default, PartialEq, Eq, Builder)]
+    pub struct Foo {
+        #[builder(mandatory)]
+        bar: Option<String>,
+    }
+    
+    let foo = Foo::builder().bar(Some("Hello world!".to_string())).build();
+}
