+++ const_typed_builder_test/compile_fail/simple_mandatory.rs
@@ -1,0 +1,11 @@
+use const_typed_builder::Builder;
+
+fn main() {
+    #[derive(Debug, Default, PartialEq, Eq, Builder)]
+    pub struct Foo {
+        bar: String,
+    }
+    
+    let foo = Foo::builder()
+        .build();
+}
