+++ const_typed_builder_test/compile_fail/simple_mandatory.stderr
@@ -1,0 +1,15 @@
+error[E0599]: no method named `build` found for struct `FooBuilder<false>` in the current scope
+  --> ./compile_fail/simple_mandatory.rs:10:10
+   |
+4  |       #[derive(Debug, Default, PartialEq, Eq, Builder)]
+   |                                               ------- method `build` not found for this struct
+...
+9  |       let foo = Foo::builder()
+   |  _______________-
+10 | |         .build();
+   | |         -^^^^^ method not found in `FooBuilder<false>`
+   | |_________|
+   |
+   |
+   = note: the method was found for
+           - `FooBuilder<true>`
